Move tests to provider package
This patch moves the tests from the `tests` directory to the package of
the code that they test.
