+++ provider/group_membership_resource_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package tests
+package provider
 
 import (
 	"context"
@@ -23,13 +23,14 @@ import (
 	"strings"
 	"time"
 
-	. "github.com/onsi/ginkgo"                         // nolint
-	. "github.com/onsi/gomega"                         // nolint
-	. "github.com/onsi/gomega/ghttp"                   // nolint
-	. "github.com/openshift-online/ocm-sdk-go/testing" // nolint
+	. "github.com/onsi/ginkgo"                                     // nolint
+	. "github.com/onsi/gomega"                                     // nolint
+	. "github.com/onsi/gomega/ghttp"                               // nolint
+	. "github.com/openshift-online/ocm-sdk-go/testing"             // nolint
+	. "github.com/openshift-online/terraform-provider-ocm/testing" // nolint
 )
 
-var _ = Describe("Machine pool creation", func() {
+var _ = Describe("Group membership creation", func() {
 	var ctx context.Context
 	var server *Server
 	var ca string
@@ -45,9 +46,9 @@ var _ = Describe("Machine pool creation", func() {
 		// Start the server:
 		server, ca = MakeTCPTLSServer()
 
-		// The first thing that the provider will do for any operation on identity providers
-		// is check that the cluster is ready, so we always need to prepare the server to
-		// respond to that:
+		// The first thing that the provider will do for any operation on grups is check
+		// that the cluster is ready, so we always need to prepare the server to respond to
+		// that:
 		server.AppendHandlers(
 			CombineHandlers(
 				VerifyRequest(http.MethodGet, "/api/clusters_mgmt/v1/clusters/123"),
@@ -69,24 +70,20 @@ var _ = Describe("Machine pool creation", func() {
 		Expect(err).ToNot(HaveOccurred())
 	})
 
-	It("Can create machine pool", func() {
+	It("Can create a group membership", func() {
 		// Prepare the server:
 		server.AppendHandlers(
 			CombineHandlers(
 				VerifyRequest(
 					http.MethodPost,
-					"/api/clusters_mgmt/v1/clusters/123/machine_pools",
+					"/api/clusters_mgmt/v1/clusters/123/groups/dedicated-admins/users",
 				),
 				VerifyJSON(`{
-				  "kind": "MachinePool",
-				  "id": "my-pool",
-				  "instance_type": "r5.xlarge",
-				  "replicas": 10
+				  "kind": "User",
+				  "id": "my-admin"
 				}`),
 				RespondWithJSON(http.StatusOK, `{
-				  "id": "my-pool",
-				  "instance_type": "r5.xlarge",
-				  "replicas": 10
+				  "id": "my-admin"
 				}`),
 			),
 		)
@@ -109,13 +106,12 @@ var _ = Describe("Machine pool creation", func() {
 				  trusted_cas = file("{{ .CA }}")
 				}
 
-				resource "ocm_machine_pool" "my_pool" {
-				  cluster      = "123"
-				  name         = "my-pool"
-				  machine_type = "r5.xlarge"
-				  replicas     = 10
+				resource "ocm_group_membership" "my_membership" {
+				  cluster   = "123"
+				  group     = "dedicated-admins"
+				  user      = "my-admin"
 				}
-					`,
+				`,
 				"URL", server.URL(),
 				"Token", token,
 				"CA", strings.ReplaceAll(ca, "\\", "/"),
@@ -124,11 +120,10 @@ var _ = Describe("Machine pool creation", func() {
 		Expect(result.ExitCode()).To(BeZero())
 
 		// Check the state:
-		resource := result.Resource("ocm_machine_pool", "my_pool")
+		resource := result.Resource("ocm_group_membership", "my_membership")
 		Expect(resource).To(MatchJQ(".attributes.cluster", "123"))
-		Expect(resource).To(MatchJQ(".attributes.id", "my-pool"))
-		Expect(resource).To(MatchJQ(".attributes.name", "my-pool"))
-		Expect(resource).To(MatchJQ(".attributes.machine_type", "r5.xlarge"))
-		Expect(resource).To(MatchJQ(".attributes.replicas", 10.0))
+		Expect(resource).To(MatchJQ(".attributes.group", "dedicated-admins"))
+		Expect(resource).To(MatchJQ(".attributes.id", "my-admin"))
+		Expect(resource).To(MatchJQ(".attributes.user", "my-admin"))
 	})
 })
