+++ provider/versions_data_source_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package tests
+package provider
 
 import (
 	"context"
@@ -23,13 +23,14 @@ import (
 	"strings"
 	"time"
 
-	. "github.com/onsi/ginkgo"                         // nolint
-	. "github.com/onsi/gomega"                         // nolint
-	. "github.com/onsi/gomega/ghttp"                   // nolint
-	. "github.com/openshift-online/ocm-sdk-go/testing" // nolint
+	. "github.com/onsi/ginkgo"                                     // nolint
+	. "github.com/onsi/gomega"                                     // nolint
+	. "github.com/onsi/gomega/ghttp"                               // nolint
+	. "github.com/openshift-online/ocm-sdk-go/testing"             // nolint
+	. "github.com/openshift-online/terraform-provider-ocm/testing" // nolint
 )
 
-var _ = Describe("Groups data source", func() {
+var _ = Describe("Versions data source", func() {
 	var ctx context.Context
 	var server *Server
 	var ca string
@@ -55,18 +56,24 @@ var _ = Describe("Groups data source", func() {
 		Expect(err).ToNot(HaveOccurred())
 	})
 
-	It("Can list groups", func() {
+	It("Can list versions", func() {
 		// Prepare the server:
 		server.AppendHandlers(
 			CombineHandlers(
-				VerifyRequest(http.MethodGet, "/api/clusters_mgmt/v1/clusters/123/groups"),
+				VerifyRequest(http.MethodGet, "/api/clusters_mgmt/v1/versions"),
+				VerifyFormKV("search", "enabled = 't'"),
 				RespondWithJSON(http.StatusOK, `{
 				  "page": 1,
-				  "size": 1,
-				  "total": 1,
+				  "size": 2,
+				  "total": 2,
 				  "items": [
 				    {
-				      "id": "dedicated-admins"
+				      "id": "openshift-v4.8.1",
+				      "raw_id": "4.8.1"
+				    },
+				    {
+				      "id": "openshift-v4.8.2",
+				      "raw_id": "4.8.2"
 				    }
 				  ]
 				}`),
@@ -91,8 +98,7 @@ var _ = Describe("Groups data source", func() {
 				  trusted_cas = file("{{ .CA }}")
 				}
 
-				data "ocm_groups" "my_groups" {
-				  cluster = "123"
+				data "ocm_versions" "my_versions" {
 				}
 				`,
 				"URL", server.URL(),
@@ -103,9 +109,11 @@ var _ = Describe("Groups data source", func() {
 		Expect(result.ExitCode()).To(BeZero())
 
 		// Check the state:
-		resource := result.Resource("ocm_groups", "my_groups")
-		Expect(resource).To(MatchJQ(`.attributes.items |length`, 1))
-		Expect(resource).To(MatchJQ(`.attributes.items[0].id`, "dedicated-admins"))
-		Expect(resource).To(MatchJQ(`.attributes.items[0].name`, "dedicated-admins"))
+		resource := result.Resource("ocm_versions", "my_versions")
+		Expect(resource).To(MatchJQ(`.attributes.items | length`, 2))
+		Expect(resource).To(MatchJQ(`.attributes.items[0].id`, "openshift-v4.8.1"))
+		Expect(resource).To(MatchJQ(`.attributes.items[0].name`, "4.8.1"))
+		Expect(resource).To(MatchJQ(`.attributes.items[1].id`, "openshift-v4.8.2"))
+		Expect(resource).To(MatchJQ(`.attributes.items[1].name`, "4.8.2"))
 	})
 })
